macros: introduce `build_field_mapping`
Move the logic for building a field mapping (which is used by the
building of format strings in `suggestion` annotations) into a helper
function.
